Ensure that planes are never on the edge
Previously, I had planes start on the edge of the radar and move off
when updateRadarWin() was called.

This is a bad idea, because it means you will never see, say, a plane on
an airport when it's taking off. It's also a special-case, since a plane
can never be on the edge again anyway.

So this commit starts the plane at (1,1)--this will soon be changed
because planes will be randomly generated anyway--and updates a few
comments and stuff.