Add delay tree
This is the beginnings of how "at beacon 0", etc. is going to work.

The delay tree is a separate tree capable of being attached to any node
of the existing tree via a function call. This means that I can reuse
the same tree for every single delayed command. The tree is only
"summoned" when it's needed.