Implement insertion sort on planesSorted
Now it's... actually sorted.

I'll use this commit message to explain why this is an array of ints,
not Plane pointers.

realloc is not guaranteed to return the same area of memory, even when
it's being called to generate an array of a smaller size.

Therefore, if I were to make planesSorted an array of pointers, it would
have to be rebuilt each time the planes array was changed. This is,
naturally, a bad thing(tm), since the whole point of caching the sort
result was to improve performance.

Anyway, this commit now actually sorts the planesSorted array.